filter: fix potential issue with output suffix
gz suffix would be added to a file in case compression is enabled.
However, that suffix was not checked when reading from compressed file.
Only file base name was used for output, resulting in gz suffix even
when compression was not enabled. To make it worse, gz would still be
appended if output was compressed, leading to silly file.pcap.gz.gz
name.

This patch sanitizes filename of all extentions, potentially sacrificing
strange naming schemes in process. But results in cleaner output. It
also improves logging as gz suffix is added by library, not application.
So application would potentially log a wrong output file name.